适配不依赖spring boot starter web的程序
