+++ log-helper-starter/log-helper-spring-boot-starter-common/src/main/java/com/redick/starter/configure/LogHelperAutoConfiguration.java
@@ -19,20 +19,8 @@ package com.redick.starter.configure;
 
 import com.redick.AroundLogHandler;
 import com.redick.banner.LogHelperBanner;
-import com.redick.starter.interceptor.SpringWebMvcInterceptor;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.format.FormatterRegistry;
-import org.springframework.http.converter.HttpMessageConverter;
-import org.springframework.validation.MessageCodesResolver;
-import org.springframework.validation.Validator;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.HandlerMethodReturnValueHandler;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-import org.springframework.web.servlet.config.annotation.*;
-
-import java.util.List;
 
 /**
  * @author Redick01
@@ -49,86 +37,5 @@ public class LogHelperAutoConfiguration {
     @Bean
     public AroundLogHandler aroundLogHandler() {
         return new AroundLogHandler();
-    }
-
-    @Bean
-    @ConditionalOnClass(name = "org.springframework.web.servlet.DispatcherServlet")
-    public WebMvcConfigurer webMvcInterceptor() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addInterceptors(InterceptorRegistry registry) {
-                registry.addInterceptor(new SpringWebMvcInterceptor());
-            }
-
-            @Override
-            public void configurePathMatch(PathMatchConfigurer configurer) {
-            }
-
-            @Override
-            public void configureContentNegotiation(ContentNegotiationConfigurer configurer) {
-            }
-
-            @Override
-            public void configureAsyncSupport(AsyncSupportConfigurer configurer) {
-            }
-
-            @Override
-            public void configureDefaultServletHandling(DefaultServletHandlerConfigurer configurer) {
-            }
-
-            @Override
-            public void addFormatters(FormatterRegistry registry) {
-            }
-
-            @Override
-            public void addResourceHandlers(ResourceHandlerRegistry registry) {
-            }
-
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-            }
-
-            @Override
-            public void addViewControllers(ViewControllerRegistry registry) {
-            }
-
-            @Override
-            public void configureViewResolvers(ViewResolverRegistry registry) {
-            }
-
-            @Override
-            public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
-            }
-
-            @Override
-            public void addReturnValueHandlers(List<HandlerMethodReturnValueHandler> handlers) {
-            }
-
-            @Override
-            public void configureMessageConverters(List<HttpMessageConverter<?>> converters) {
-            }
-
-            @Override
-            public void extendMessageConverters(List<HttpMessageConverter<?>> converters) {
-            }
-
-            @Override
-            public void configureHandlerExceptionResolvers(List<HandlerExceptionResolver> resolvers) {
-            }
-
-            @Override
-            public void extendHandlerExceptionResolvers(List<HandlerExceptionResolver> resolvers) {
-            }
-
-            @Override
-            public Validator getValidator() {
-                return null;
-            }
-
-            @Override
-            public MessageCodesResolver getMessageCodesResolver() {
-                return null;
-            }
-        };
     }
 }
