+++ log-helper-starter/log-helper-spring-boot-starter-common/src/main/resources/META-INF/spring.factories
@@ -4,4 +4,5 @@ org.springframework.context.ApplicationContextInitializer=\
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.redick.starter.configure.SpringRedisAutoConfiguration,\
   com.redick.starter.configure.JedisAutoConfiguration,\
-  com.redick.starter.configure.LettuceAutoConfiguration
+  com.redick.starter.configure.LettuceAutoConfiguration,\
+  com.redick.starter.configure.WebMvcConfiguration
